ckpt 342: B1 done: NOVIG-V3 signer (BouncyCastle) verified against all 30 Novig vectors + published signatures; NovigSignedClient; StreamBooks + NovigStream (websocket book channel, token-budgeted chunked subscribes) in data module. data 75 tests green. Secret-scan false positive on runtime-generated test PEMs resolved with the scanner's documented FAKE marker. v0.4.0 release run 36100439547 in progress
next: Confirm v0.4.0 release + BUILDLOG + send link; then B2 (app Settings Novig key setup, Keystore P-256 read key) + B3 (wire NovigStream into Scanner)
tests: all 1 fast checks green

+++ data/src/main/kotlin/com/tjshea/vigilant/data/novig/stream/NovigStream.kt
@@ -67,7 +67,7 @@ class NovigStream(
         _state.value = StreamState.Connecting
         val signed = signer.signedRequest("GET", "/v3/ws")
         val request = signed.newBuilder().url(wsUrl).build()
-        tokens = STREAM_CAPACITY - UPGRADE_COST
+        tokens = (STREAM_CAPACITY - UPGRADE_COST).toDouble()
         tokensAt = clock()
         socket = http.newWebSocket(request, Listener())
     }
